+++ R/zzz.R
@@ -1,25 +1,9 @@
 #' @import shiny
 #' @import tfrmt
-#' @importFrom shinyjs useShinyjs enable disable addClass removeClass toggleState show hide toggle onclick toggleClass
 #' @import dplyr
-#' @import stringr
-#' @import purrr
-#' @importFrom DT datatable DTOutput renderDT
-#' @import gt
-#' @import sortable
-#' @importFrom rlang quo_is_missing
+#' @importFrom rlang quo sym quo_is_missing `:=` `!!` as_label is_empty
 #' @import bslib
-#' @import shinyWidgets
-#' @import forcats
-#' @importFrom rio import
-#' @importFrom shinycssloaders withSpinner
 #' @import webshot2
-#' @import shinyAce
-#' @importFrom stats setNames
-#' @importFrom rlang `:=` `!!`
-#' @importFrom shinyFeedback useShinyFeedback feedbackDanger
-#' @importFrom fontawesome fa_i
-#' @importFrom utils getFromNamespace zip
 NULL
 
 globalVariables(".")