+++ tests/testthat/module_examples/mod_datamapping/app.R
@@ -18,7 +18,7 @@ shinyApp(
     includeCSS(system.file("www","styles.css", package = "tfrmtbuilder")),
     html_dependency_pretty(),
     useShinyjs(),
-    useShinyFeedback(),
+    shinyFeedback::useShinyFeedback(),
     datamapping_ui("mappings")
   ),
   function(input,output,session){